+++ CHANGELOG.md
@@ -2,11 +2,12 @@
 
 All notable changes to this project will be documented in this file.
 
-## [Unreleased]
+## [v7.3.0] - 2024-04-30
 
 - **Feature**: Allow excluding files and folders from indexing with regex pattern ([#1015](https://github.com/pulsejet/memories/pull/1015))
 - **Feature**: Allow forcing software-based transpose for transcoding ([#794](https://github.com/pulsejet/memories/issues/794))
 - **Note**: Using Imaginary is now discouraged for now due to multiple bugs. The admin panel will show a warning if Imaginary is enabled.
+- Various bug fixes.
 
 ## [v7.2.0] - 2024-04-05